+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Head
+# 7.0.2
 
 - Fixed: `at-rule-blacklist`, `at-rule-whitelist`, `comment-word-blacklist`, `selector-attribute-operator-blacklist`, `selector-attribute-operator-whitelist` now accept array as first option.
 - Fixed: `unit-*` rules now ignore CSS hacks.